# server/swagger/swagger.yaml
definitions:
...
  api.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
  api.Meta:
    properties:
      error:
        type: string
      msg:
        description: Status int64  `json:"status"`
        type: string
    type: object
...
  api.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/api.Meta'
    type: object
...
  api.UpdateProjectAssHostReq:
    properties:
      hid:
        items:
          type: integer
        type: array
      pid:
        type: integer
    required:
    - hid
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/project/association:
    put:
      description: 服务器ID[多选]
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关联传入参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.UpdateProjectAssHostReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 关联服务器
      tags:
      - 项目相关
  /api/v1/project/delete:
    delete:
      description: 返回success
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除project所需参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 删除项目
      tags:
      - 项目相关
...
  /api/v1/project/getProject:
    get:
      description: 返回指定项目
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: name
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 获取项目
      tags:
      - 项目相关
  /api/v1/project/getSelfProject:
    get:
      description: 返回自身所属项目
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 获取自身所属项目
      tags:
      - 项目相关
  /api/v1/project/update:
    post:
      description: 返回新增/修改的指定项目
      parameters:
      - description: 格式为：Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: group_id
        required: true
        type: integer
      - description: 修改才需要传，没有传算新增
        in: formData
        name: id
        type: integer
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: status
        required: true
        type: integer
      - in: formData
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"data":{},"meta":{msg":"Success"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "403":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: '{"data":{}, "meta":{"msg":"错误信息", "error":"错误格式输出(如存在)"}}'
          schema:
            $ref: '#/definitions/api.Response'
      summary: 新增/修改项目
      tags:
      - 项目相关
...
swagger: "2.0"